Rename Subgraphs.h to Model.h
